Avoid wrapping children
This changes the API so that by default the FocusTrap doesn't wrap
the children and therefore there must only be a single child element
under HotKeys. In such a case there is no tabIndex manipulation
unless used explicitly.

In order to wrap the contents one needs to explicitly define the
component property as follows:

  <HotKeys>
     <p>This is NOT going to get wrapped with a div</p>
  </HotKeys>

  <HotKeys component="div">
     <p>This is going to get wrapped around with a div</p>
  </HotKeys>