Correcly handle when modifier key's keyup events are missed
When the user is performing actions that may result in a keyup not being seen (such as alt+tab-ing away), when the user returns, react-hotkeys still "thought" one or more of the modifier keys were held down. This update uses the modifier key attribute flags on every event to interpolate when a modifier key has been released, but not detected.